fix: resolve 9 critical compilation errors and add missing core providers
**Critical Fixes Applied (9 errors → 0 errors)**:

1. core/providers.dart - Create missing global Isar provider
   - Centralizes database instance initialization
   - Supports all 8 DTO collection schemas
   - Enables Phase 1 migration to Supabase (1-line change)

2. pubspec.yaml - Add intl package dependency
   - Enables DateFormat functionality for dashboard

3. DataSharingNotifier - Fix provider return type
   - Add explicit return type to sharedDataRepository provider
   - Eliminates nullable type safety error

4. IsarSharedDataRepository - Fix Isar API compatibility
   - Update legacy .findAll() to modern .where().findAll()
   - Maintains Isar version compatibility

5. OnboardingScreen - Fix nullable parameter types
   - Add null-coalescing operators for userId and onComplete

6. UserBadgeDto - Add missing copyWith() method
   - Enables BadgeRepository mutation operations
   - Includes all 9 fields with proper null-safety

7. DataSharingAggregator - Fix num to double type cast
   - Add explicit .toDouble() conversion

8. DataSharingRepository - Fix List<dynamic> type cast
   - Use proper .cast<SymptomLog>() for type safety

9. DataSharingScreen - Fix ambiguous provider imports
   - Remove duplicate dataSharingNotifierProvider definition
   - Use single authoritative provider source

**Result**: 51 total issues (0 errors, all info/warning level)

+++ lib/features/data_sharing/domain/usecases/data_sharing_aggregator.dart
@@ -17,7 +17,7 @@ class DataSharingAggregator {
     final startWeight = sortedWeights.first.weightKg;
     final endWeight = sortedWeights.last.weightKg;
     final change = endWeight - startWeight;
-    final changePercentage = startWeight > 0 ? (change / startWeight) * 100 : 0;
+    final changePercentage = (startWeight > 0 ? (change / startWeight) * 100 : 0).toDouble();
 
     return WeightTrend(
       startWeight: startWeight,
